major refactoring and decupling of frontend and logic
