+++ tests/test_configs.py
@@ -1,8 +1,10 @@
 """Test existing config files for schema."""
 
+from pathlib import Path
+
 import pytest
 import yaml
-from pathlib import Path
+
 from openandroidinstaller.installer_config import validate_config
